:truck: api: move files to src
