+++ api/src/certification/session/domain/usecases/validate-live-alert.js
@@ -1,6 +1,15 @@
 import { NotFoundError } from '../../../../../lib/domain/errors.js';
+import { CertificationIssueReport, CertificationIssueReportCategory } from '../../../../../lib/domain/models/index.js';
 
-export const validateLiveAlert = async ({ userId, sessionId, certificationChallengeLiveAlertRepository }) => {
+export const validateLiveAlert = async ({
+  userId,
+  sessionId,
+  subcategory,
+  certificationChallengeLiveAlertRepository,
+  assessmentRepository,
+  issueReportCategoryRepository,
+  certificationIssueReportRepository,
+}) => {
   const certificationChallengeLiveAlert =
     await certificationChallengeLiveAlertRepository.getOngoingBySessionIdAndUserId({
       sessionId,
