+++ api/src/certification/shared/infrastructure/repositories/issue-report-category-repository.js
@@ -1,4 +1,4 @@
-import { knex } from '../../../db/knex-database-connection.js';
+import { knex } from '../../../../../db/knex-database-connection.js';
 
 const get = async function ({ name }) {
   return knex('issue-report-categories').where({ name }).first();
